+++ CHANGELOG.md
@@ -29,3 +29,6 @@ v0.21 + 1
 * The git_clone_options struct now provides repository_cb and
   repository_cb_payload to allow the user to create a repository with
   custom options.
+
+* git_clone_into and git_clone_local_into have been removed from the
+  public API in favour of git_clone callbacks